property list is almost done, just styling
