+++ client/package.json
@@ -9,6 +9,7 @@
     "@testing-library/user-event": "^7.1.2",
     "graphql": "^15.3.0",
     "leaflet": "^1.7.1",
+    "node-sass": "^5.0.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-leaflet": "^3.0.2",
