+++ client/src/Components/PropertyList/index.js
@@ -1,2 +1,1 @@
-export { Map } from './Map'
 export { PropertyList } from './PropertyList'
